pyproject.toml
[tool.poetry]
name = "fastio-cli"
version = "0.1.0"
description = ""
authors = ["fastio.dev <opensource@fastio.dev>"]

[tool.poetry.scripts]
fastio = "fastio_cli.cli:cli"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.1"
pydantic = "^1.8.2"
Jinja2 = "^3.0.1"
GitPython = "^3.1.24"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.12.1"
pytest = "^6.2.5"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
